requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.6
deepdiff==7.0.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
emoji==2.12.1
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
filetype==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langdetect==1.0.9
langsmith==0.1.63
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.30.4
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pydantic==2.7.2
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.3
Pygments==2.18.0
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
rapidfuzz==3.9.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
ujson==5.10.0
unstructured==0.14.2
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.30.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
